Create Transcribing_DNA_to_RNA
Tutorial 
We start by creating a DNAtoRNA class that contains our main method.
Inside the main method, we create a string variable called dnaSequence and assign it a DNA sequence value. For the purposes of this example, I've chosen a simple 20-base sequence, but you can use any DNA sequence you like.
We then transcribe the DNA sequence to RNA by using the replace method of the string class. This method takes two arguments: the character to be replaced ('T' in our case), and the character to replace it with ('U').
We store the resulting RNA sequence in a new string variable called rnaSequence.
Finally, we display both the DNA and RNA sequences to the screen using the println method of the System.out object.