Create List of Amino Acid abbreviations
Inside the main method, we define an array called aminoAcids that contains the list of amino acid abbreviations.
We use the Scanner class to get the user's input for the amino acid abbreviation.
We define a boolean variable called found and set it to false.
We use a for-each loop to iterate through each amino acid abbreviation in the aminoAcids array.
Inside the loop, we check whether the current amino acid abbreviation (acid) is equal to the user's input amino acid abbreviation (inputAcid) using the equals method. If they are equal, we set found to true and break out of the loop.
After the loop, we use an if-else statement to print whether the amino acid was found or not. If found is true, we print "Amino acid found". Otherwise, we print "Amino acid NOT found".