Create Read in a whole genome (in FASTA format)
We first define the filename of the input file as a String variable.
We create a StringBuilder object to store the genome sequence as we read it in.
We define a String variable "line" to store each line of the input file as we read it in.
We use a try-with-resources statement to open the input file as a BufferedReader.
We loop through each line in the file, appending the nucleotides to the StringBuilder object. We skip the header line (which starts with ">").
If there is an error reading the file, we print an error message and exit the program.
We convert the StringBuilder object to a String and print the length of the genome sequence to the console.
To use this program, save it as a Java file (e.g. "GenomeReader.java") and compile it with the command "javac GenomeReader.java". Then, run the program with the command "java GenomeReader". The program will read in the genome sequence from the "genome.txt" file and print its length to the console.

You can modify this program to perform other tasks on the genome sequence, such as searching for specific sequences or analyzing the GC content.