Create DNA Sequence Analyzer
Project Title: DNA Sequence Analyzer

Project Description:
The DNA Sequence Analyzer project aims to develop a Java-based tool that can analyze and manipulate DNA sequences. The program will allow the user to input a DNA sequence in FASTA format, and perform various analyses on it. Some of the functionalities that the program will provide are:

Reverse Complement: The program will provide the reverse complement of the input sequence. This is useful for analyzing DNA sequences that are read in the reverse direction.

Translation: The program will translate the input DNA sequence to its corresponding amino acid sequence. This functionality will help in identifying open reading frames in a DNA sequence.

Codon Usage: The program will analyze the codon usage in the input sequence. This analysis will help in understanding the genetic code and the possible protein sequences that can be synthesized from the given DNA sequence.

GC Content: The program will calculate the GC content of the input DNA sequence. This is useful for identifying regions of DNA that are more stable and resistant to mutation.

BLAST: The program will provide basic BLAST functionality to allow the user to search for similar sequences in a database.

Tools and Technologies:

Java 8 or higher
BioJava library for bioinformatics calculations
BLAST+ software for BLAST searches
Project Deliverables:

Java code for the DNA Sequence Analyzer
User manual and documentation for the program
Sample input DNA sequences and expected outputs for each of the functionalities provided by the program
Optional Features:

GUI implementation for the program
Integration with external databases such as NCBI or UniProt
Support for multiple sequence alignment and phylogenetic analysis